+++ dashboard/resources.py
@@ -5,7 +5,8 @@ from typing import List
 from starlette.requests import Request
 
 from dashboard import enums
-from dashboard.biz_models.datamanager import DataSet, EvaluationPlan
+from dashboard.biz_models import DataSet, EvaluationPlan
+from dashboard.biz_models import LabelPage
 from dashboard.constants import BASE_DIR
 
 # from dashboard.models import Evaluation
@@ -14,11 +15,12 @@ from dashboard.models import Cat  # EvaluationPlan,; Evaluation,
 from dashboard.models import Category  # EvaluationPlan,; Evaluation,
 from dashboard.models import Config  # Evaluation,
 from dashboard.models import Dog1  # EvaluationPlan,; Evaluation,
-from dashboard.models import LabelPage, Log  # EvaluationPlan,; Evaluation,
+from dashboard.models import Log
 from dashboard.models import Permission as PermissionModel
-from dashboard.models import Product, Record  # EvaluationPlan,; Evaluation,
+from dashboard.models import Product
 from dashboard.models import Resource as ResourceModel
 from dashboard.models import Role as RoleModel
+from dashboard.biz_models.eval_model import Record
 
 # from dashboard.models import Sponsor
 from dashboard.providers import import_export_provider
@@ -403,7 +405,7 @@ class DataManager(Dropdown):
                 name="eval_type",
                 label="评分方式",
                 display=displays.InputOnly(),
-                input_=inputs.RadioEnum(enums.ScoreWayType, default=enums.ScoreWayType.system),
+                input_=inputs.RadioEnum(enums.EvaluationType, default=enums.EvaluationType.auto_ai_critique),
             ),
         ]
