+++ dashboard/templates/record/create_eval.html
@@ -6,7 +6,7 @@
             <label class="form-label required">{{ _('Evaluation Plan') }}</label>
             <select required class="form-select" id="eval_plan">
                 {% for item in eval_plans %}
-                <option value="{{ item.id }}">{{ item.plan_name }}</option>
+                <option value="{{ item.id }}">{{ item.name }}</option>
                 {% endfor %}
             </select>
         </div>
